kernel: Only expose su when daemon is running
It has been claimed that the PG implementation of 'su' has security
vulnerabilities even when disabled. Unfortunately, the people that
find these vulnerabilities often like to keep them private so they
can profit from exploits while leaving users exposed to malicious
hackers.

In order to reduce the attack surface for vulnerabilites, it is
therefore necessary to make 'su' completely inaccessible when it
is not in use (except by the root and system users).

Change-Id: Ia7d50ba46c3d932c2b0ca5fc8e9ec69ec9045f85
https://review.lineageos.org/#/c/170648/

kernel: Fix potential refcount leak in su check

Change-Id: I3d241ae805ba708c18bccfd5e5d6cdcc8a5bc1c8
https://review.lineageos.org/#/c/173619/