Fix build for drivers/video/msm/mdss/mdss_hdmi_cec.c
The fix for CVE-2017-9719, a backport from kernel 3.18 in
commit d108c84f3c20a832a60f5b5773e8b96a12bba0b2, involves
definitions MAX_CEC_FRAME_SIZE and MAX_OPERAND_SIZE into
file mdss_hdmi_cec.c
In kernel 3.18, this is defined in file mdss_core_cec.h,
which does not exist in our 3.4 kernel - this file is part
of a major code redesign in 3.18, which cannot simply be
backported. The missing definitions are therefore added to
mdss_hdmi_cec.h in our backport to 3.4

In order to be able to revert this easily, this is done in
this separate commit.

Change-Id: I8f5a770b29cdafeb3f91b58a3adcff5a59ebc956